加点注释

+++ sample/src/main/java/com/norman/android/hdrsample/util/TimeUtil.java
@@ -8,22 +8,48 @@ public class TimeUtil {
 
     private static final long SECOND_TO_MILL_SCALE = 1000L;
 
+    /**
+     * 微妙转毫秒
+     * @param timeUs
+     * @return
+     */
     public static long microToMill(long timeUs){
         return TimeUnit.MICROSECONDS.toMillis(timeUs);
     }
 
+    /**
+     * 纳秒转微妙
+     * @param timeNs
+     * @return
+     */
     public static long nanoToMicro(long timeNs){
         return TimeUnit.NANOSECONDS.toMicros(timeNs);
     }
 
+    /**
+     * 微妙转秒
+     * @param timeUs
+     * @return
+     */
     public static float microToSecond(long timeUs){
-        return (float) (timeUs*1.0/ SECOND_TO_MICRO_SCALE);
+        return timeUs*1.0f/ SECOND_TO_MICRO_SCALE;
     }
+
+    /**
+     * 秒转微妙
+     * @param second
+     * @return
+     */
 
     public static long secondToMicro(float second){
         return (long) (second* SECOND_TO_MICRO_SCALE);
     }
 
+    /**
+     * 秒转毫秒
+     * @param second
+     * @return
+     */
     public static long secondToMill(float second){
         return (long) (second* SECOND_TO_MILL_SCALE);
     }
